General code cleanup from initial
