+++ pkg/user/profile.go
@@ -3,7 +3,6 @@ package user
 import (
 	"encoding/json"
 	"fmt"
-	"math"
 )
 
 // Profile is Model with device, location, profile information
@@ -57,26 +56,4 @@ func (u *Profile) String() string {
 
 	// Return String
 	return string(msgBytes)
-}
-
-// Update takes json and updates status/direction
-func (u *Profile) Update(direction float64) {
-
-	// Set New Data
-	u.Direction = Round(direction, .5, 2)
-}
-
-// Round converts a number to be rounded
-func Round(val float64, roundOn float64, places int) (newVal float64) {
-	var round float64
-	pow := math.Pow(10, float64(places))
-	digit := pow * val
-	_, div := math.Modf(digit)
-	if div >= roundOn {
-		round = math.Ceil(digit)
-	} else {
-		round = math.Floor(digit)
-	}
-	newVal = round / pow
-	return
 }
