+++ pkg/util/p2p.go
@@ -1,4 +1,4 @@
-package host
+package util
 
 import (
 	"context"
